donate: merge membership and one-time bullets
The donation forms are now side-by-side, which looks better, but the
membership bullet point referred to a form which was part of the
following bullet point.

This also removes the mention of "so far we have been fortunate" to
always be able to pay developers because it is no longer true.

"recurring payment" and "one-time" are no longer emphasized.