Introduce Markdown and HTML classes for bug discovery
The HTML class should encapsulate all HTML code. This should make it
easier to discover unintentional misuse of markdown, i.e. trying to
markdown-format HTML code or already formatted markdown code.
It is designed in such a way that strings (or values convertible to
strings) can be appended to it, yet the result will always be tainted
by bearing a HTML class.

The Markdown class is a shallow wrapper around HTML to give more
sensible error messages in case of markdown-formatting an already
markdown-formatted string. Further checks may be added in the future.

The actual checks are still disabled in this commit.