Annotate all template tags with their respective types, fix double <p>
substitute_html now checks if template tags starting with html__, str__
or md__ are substituted with values of type HTML, str, Markdown
respecively. This allows to indicate the intended kind of substitution
expected in the respective part of the template. In turn, this makes
it easier for human beings to track down missing or duplicate <p> tags,
and prevents future unintentional misuse of the template.

Consequently, this fixes some double <p> and other invalid HTML
constructs as well.