Dynamically update <body> padding-top to match navigation bar height
This introduces some Javascript code that queries the height of the
navigation bar each time the window is resized, which should be the
only times the navigation bar changes its wrapping state (either from
wrapped to normal or vice versa).

With the up-to-date knowledge on the navigation bar height, it then
reconfigures the padding-top on the body element to effectively shift
all content downwards to exactly match the navigation bar height.

To gracefully degrade when Javascript is not supported, this changes
the default navigation bar class to navbar-static-top, which behaves as
if attached to and directly above all content (so it actually occupies
space and scrolls with the page). Then, in Javascript, this class is
substituted with navbar-fixed-top (which affixes the navigation bar to
the window top, where it remains visible when navigating the page) and
the body padding-top workaround applied.