Catch anchor clicks and hashchange to account for navigation bar height
This adds additional Javascript code to catch clicks on links
referring to anchors / "hashes" on the page, and apply the navigation
bar height offset to the scroll position. By this means we no longer
need to rely on known-excessive offsets with CSS tricks.

Graceful degradation when Javascript is not supported is guaranteed by
the existing navigation bar display logic: without Javascript, the
navigation bar is already displayed in navbar-static-top mode (as if
attached to the content) so offsetting is not required.