Fix single-pixel jump when navigation bar item becomes active
Without active navigation bar item, the menu bar was 1px less tall than
when an item was active, since the additional single-pixel border
increases the height of the element which is not accounted for by a
decreased padding. This commit fixes this behaviour by decreasing the
single-pixel padding to none to account for the single-pixel border.