Fix page jump while loading
This change fixes the page jumping a little while the page is being
loaded in Chrome. Here, the page was first being rendered with the
navbar-static-top navigation bar, which has margin and border
properties set. This resulted in an improperly calculated navigation
bar height, which in turn causes the page to jump when switching to
the navbar-fixed-top navigation bar. This is fixed by also taking into
account the vertical margin and border of the navbar-static-top.