"About" menu: Remove "People" section, migrated to Wiki
https://wiki.freenetproject.org/index.php?title=People&oldid=3187
https://wiki.freenetproject.org/index.php?title=Main_Page&diff=3188&oldid=3008

Reasons for migration:

- Our contributors change very frequently, it will be easier to maintain
the list in something which can be as comfortable updated as a Wiki

- The website should be focused on providing a good experience for
users. Users won't care *who* wrote the software though, they mostly
will care about what it can do for them.

- The whole of the "About" menu is actually rendered into one HTML page,
including the "People" section. First time visitors are very likely to
be interested in what Freenet is, so they are likely to run into the
"About" page. Hence it shouldn't be too long.