Homepage: Simplify slogan and mission
The main goal of this PR is to reduce the amount of text in the view
focus of the first impression of the user.
To do so, it chooses more simple English at the slogan and mission,
which additionally makes them more easy to understand.

About the slogan:
I usually program at a public place where people with laptops are rather
unusual. That has made hundreds of them ask me what I'm doing with my
laptop. Over the years I've tried various approaches of explaining what
Freenet is. Those folks are unrelated to IT, so they're a good sample.
The approach which people seem to get the quickest is telling them that
Freenet is a darknet. The word really has become mainstream, people
understand it. So it needs to be part of the slogan.
Beyond that, I consider "Leap" and "Surveillance" as less common English
words than "Join". So I've replaced the old slogan with this more simple
one.
I could have gone for just "Join the Darknet", but I think we need to
offer one key feature in our slogan. I consider our two main features
to be anonymity and censorship resistance. The choice went for only
"anonymous" to be contained in the slogan because:
Being anonymous is more apparently useful from a user's perspective
since surveillance feels more visible in daily life than censorship.
If things are censored, you won't even know this most of the time
because they are just gone. But only from the daily targeted advertising
on the regular Internet, you'll for sure know that you're not anonymous.
Everyone has heard about the NSA scandal as well. So people will first
want to have surveillance fixed before censorship, and thus "anonymous"
is the better candidate for being mentioned in the slogan.
In return, censorship is mentioned in the mission.

About the mission:
Remember: This PR wants to get shorter, easier to understand text.
So to shorten the text, I left out:
- technical details which are uninteresting to average people
("peer-to-peer")
- complex or long words ("platform", "censorship-resistant",
"communication", "privacy protections")
- Specific examples of apps on Freenet ("forums" etc.). They were
replaced with the generic Web2.0 terms of "speak, share and post".
(Those terms were sorted in the order of commonness so you read the more
easy to understand ones first.)

Notably, our two main features of anonymity and censorship resistance
are both mentioned twice with different wording:
Anonymity once in the slogan and once in the mission.
Censorship twice in the mission.
This is to burn them into the user's memory: Mentioning important
concepts in different wording twice is a good concept for making sure
people memorize them. In fact I just did that with this commit message
as well :)